generalise out to hooks
